+++ apps/meteor/client/lib/cachedCollections/index.ts
@@ -1,3 +1,4 @@
 export { PrivateCachedCollection, PublicCachedCollection } from './CachedCollection';
 export { CachedCollectionManager } from './CachedCollectionManager';
 export { pipe } from './pipe';
+export { applyQueryOptions, convertSort } from './utils';
