+++ packages/ui-contexts/src/index.ts
@@ -99,6 +99,6 @@ export { useOnLogout } from './hooks/useOnLogout';
 
 export { UploadResult } from './ServerContext';
 export { TranslationKey, TranslationLanguage } from './TranslationContext';
-export { Fields } from './UserContext';
+export { Fields, FindOptions } from './UserContext';
 
 export { SubscriptionWithRoom } from './types/SubscriptionWithRoom';
